refactor(types): create shared type contract for webview
Centralizes all type definitions used for communication between the extension host and the webview into a new `types/webview.ts` file.

This creates a single source of truth for the data contract, reducing code duplication and improving maintainability.

- Removes duplicated type definitions from `SettingsPanel.ts` and `webview/main.ts`.
- Updates all relevant source files and test files to import from the new shared location.
- Fixes all test suite compilation errors that arose from the refactoring.
