+++ CHANGELOG.md
@@ -1,16 +1,18 @@
 # Changelog
 
-## [1.6.3] - 2025-08-26
+## [1.6.3] - 2025-08-27
 
 - **Feature**: Added `.dockerignore` file
 - **Feature**: The hex code input fields for syntax colors in the Customization UI are now fully interactive. Users can type or paste a hex code, including an 8-digit code with transparency, and the color picker, alpha slider, and live preview will update in real-time.
 - **Feature**: The Customization UI now features a view selector to switch between "Syntax", "Brackets", and "JSON" settings. This organizes the panel and provides dedicated live previews for each category.
 - **Feature**: Added unit tests for the webview.
-- **Improvement**: Cached github workflow for the npm dependencies and docker image so that the pipline is significant faster.
-- **Improvement**: Docker caching for github actions.
+- **Improvement**: Cached github workflow for the npm dependencies and docker image so that the pipline is faster.
+- **Internal Refactor**: Centralized the type definitions for communication between the extension and the Customization UI into a single, shared `types/webview.ts` file. This creates a stricter and more maintainable contract, reducing code duplication and improving overall reliability.
+- **Chore**: Removed an unused import from the `WebviewManager` to improve code cleanliness.
 - **Fix**: Updated the extension's configuration schema (package.json). VS Code will now correctly validate 8-digit hex codes with alpha channels when a user edits their settings.json manually.
 - **Fix**: Corrected a CSS grid layout issue that caused the bracket color settings to be displayed in a disorganized manner.
 - **Fix**: Updated the webview unit tests to align with the new data structures, resolving a test failure caused by the recent feature addition.
+- - **Fix**: Updated the unit and integration tests (`SettingsPanel.test.ts`, `WebviewManager.test.ts`) to use the new shared type definitions, resolving all compilation errors and ensuring the test suite passes.
 
 ## [1.6.2] - 2025-08-25
